requirements.txt
# OCR Benchmark Requirements

# Azure Document Intelligence
azure-ai-documentintelligence>=1.0.0
azure-core>=1.29.0

# Google Gemini
google-genai>=0.3.0

# LlamaCloud (LlamaParse)
llama-parse>=0.4.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0

# Groq API
Groq

